requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
numpy==1.23.5
tensorflow==2.13.0
Pillow==9.5.0
opencv-python==4.11.0.86